+++ orm/bucket.go
@@ -37,14 +37,31 @@ type Bucket interface {
 	DBKey(key []byte) []byte
 	Delete(db weave.KVStore, key []byte) error
 	Get(db weave.ReadOnlyKVStore, key []byte) (Object, error)
+	// Index returns an index with given name maintained for this bucket.
+	Index(name string) (Index, error)
 	GetIndexed(db weave.ReadOnlyKVStore, name string, key []byte) ([]Object, error)
-	GetIndexedLike(db weave.ReadOnlyKVStore, name string, pattern Object) ([]Object, error)
 	Parse(key, value []byte) (Object, error)
 	Register(name string, r weave.QueryRouter)
 	Save(db weave.KVStore, model Object) error
 	Sequence(name string) Sequence
+
+	// WithIndex returns a copy of this bucket with given index. Index is
+	// maintained as a single set. This implementation is suitable for
+	// small collections.
+	// Panics if it an index with that name is already registered.
 	WithIndex(name string, indexer Indexer, unique bool) Bucket
+
+	// WithMultiKeyIndex returns a copy of this bucket with given index.
+	// Index is maintained as a single set. This implementation is suitable
+	// for small collections.
+	// Panics if it an index with that name is already registered.
 	WithMultiKeyIndex(name string, indexer MultiKeyIndexer, unique bool) Bucket
+
+	// WithNativeIndex returns a copy of this bucket with given index.
+	// Index is maintained using database native support. This
+	// implementation is suitable for big collections.
+	// Panics if it an index with that name is already registered.
+	WithNativeIndex(name string, indexer MultiKeyIndexer) Bucket
 }
 
 // bucket is a generic holder that stores data as well
@@ -60,30 +77,30 @@ type bucket struct {
 	prefix []byte
 	model  reflect.Type
 	// index is a list of indexes sorted by
-	indexes namedIndexes
+	indexes boundIndexes
 }
 
 var _ Bucket = (*bucket)(nil)
 
-type namedIndex struct {
-	Index
+type bucketBoundIndex struct {
+	idx        Index
 	publicName string
 }
 
-type namedIndexes []namedIndex
+type boundIndexes []bucketBoundIndex
 
 // Get returns the index with the given (internal/db) name, or nil if not found
-func (n namedIndexes) Get(name string) *Index {
+func (n boundIndexes) Get(name string) Index {
 	for _, ni := range n {
 		if ni.publicName == name {
-			return &ni.Index
+			return ni.idx
 		}
 	}
 	return nil
 }
 
 // Has returns true iff an index with the given name is already registered
-func (n namedIndexes) Has(name string) bool {
+func (n boundIndexes) Has(name string) bool {
 	return n.Get(name) != nil
 }
 
@@ -110,7 +127,7 @@ func (b bucket) Register(name string, r weave.QueryRouter) {
 	root := "/" + name
 	r.Register(root, b)
 	for _, ni := range b.indexes {
-		r.Register(root+"/"+ni.publicName, ni.Index)
+		r.Register(root+"/"+ni.publicName, ni.idx)
 	}
 }
 
@@ -227,8 +244,8 @@ func (b bucket) updateIndexes(db weave.KVStore, key []byte, model Object) error 
 		if err != nil {
 			return err
 		}
-		for _, idx := range b.indexes {
-			err = idx.Update(db, prev, model)
+		for _, ni := range b.indexes {
+			err = ni.idx.Update(db, prev, model)
 			if err != nil {
 				return err
 			}
@@ -240,6 +257,23 @@ func (b bucket) updateIndexes(db weave.KVStore, key []byte, model Object) error 
 // Sequence returns a Sequence by name
 func (b bucket) Sequence(name string) Sequence {
 	return NewSequence(b.name, name)
+}
+
+func (b bucket) WithNativeIndex(name string, indexer MultiKeyIndexer) Bucket {
+	if b.indexes.Has(name) {
+		panic(fmt.Sprintf("Index %s registered twice", name))
+	}
+
+	iname := b.name + "_" + name
+	idxs := append(b.indexes, bucketBoundIndex{
+		idx:        NewNativeIndex(iname, indexer),
+		publicName: name,
+	})
+	sort.Slice(idxs, func(i int, j int) bool {
+		return idxs[i].idx.Name() < idxs[j].idx.Name()
+	})
+	b.indexes = idxs
+	return b
 }
 
 // WithIndex returns a copy of this bucket with given index,
@@ -258,10 +292,18 @@ func (b bucket) WithMultiKeyIndex(name string, indexer MultiKeyIndexer, unique b
 
 	iname := b.name + "_" + name
 	add := NewMultiKeyIndex(iname, indexer, unique, b.DBKey)
-	idxs := append(b.indexes, namedIndex{Index: add, publicName: name})
-	sort.Slice(idxs, func(i int, j int) bool { return idxs[i].name < idxs[j].name })
+	idxs := append(b.indexes, bucketBoundIndex{idx: add, publicName: name})
+	sort.Slice(idxs, func(i int, j int) bool { return idxs[i].idx.Name() < idxs[j].idx.Name() })
 	b.indexes = idxs
 	return b
+}
+
+func (b bucket) Index(name string) (Index, error) {
+	idx := b.indexes.Get(name)
+	if idx == nil {
+		return nil, errors.Wrap(ErrInvalidIndex, name)
+	}
+	return idx, nil
 }
 
 // GetIndexed queries the named index for the given key
@@ -270,20 +312,7 @@ func (b bucket) GetIndexed(db weave.ReadOnlyKVStore, name string, key []byte) ([
 	if idx == nil {
 		return nil, errors.Wrap(ErrInvalidIndex, name)
 	}
-	refs, err := idx.GetAt(db, key)
-	if err != nil {
-		return nil, err
-	}
-	return b.readRefs(db, refs)
-}
-
-// GetIndexedLike queries the named index with the given pattern
-func (b bucket) GetIndexedLike(db weave.ReadOnlyKVStore, name string, pattern Object) ([]Object, error) {
-	idx := b.indexes.Get(name)
-	if idx == nil {
-		return nil, errors.Wrap(ErrInvalidIndex, name)
-	}
-	refs, err := idx.GetLike(db, pattern)
+	refs, err := consumeIteratorKeys(idx.Keys(db, key))
 	if err != nil {
 		return nil, err
 	}