+++ orm/index.go
@@ -2,12 +2,38 @@ package orm
 
 import (
 	"bytes"
+	math "math"
 
 	"github.com/iov-one/weave"
 	"github.com/iov-one/weave/errors"
 )
 
-var indPrefix = []byte("_i.")
+type Index interface {
+	// Name returns the name of this index.
+	Name() string
+
+	// Update updates the index. It should be called when any of the bucket
+	// entities has changed in the store.
+	//
+	// prev == nil means insert
+	// save == nil means delete
+	// both == nil is error
+	// if both != nil and prev.Key() != save.Key() this is an error
+	Update(db weave.KVStore, prev Object, save Object) error
+
+	// Keys returns an iteator that returns all entity keys that were
+	// indexed under given value.
+	//
+	// Values of returned iterator are always nil to optimize for a lazy
+	// loading flows and avoid loading into memory values from the database
+	// when they might not be needed.
+	Keys(db weave.ReadOnlyKVStore, value []byte) weave.Iterator
+
+	// Query handles queries from the QueryRouter.
+	Query(db weave.ReadOnlyKVStore, mod string, data []byte) ([]weave.Model, error)
+}
+
+const compactIdxPrefix = "_i."
 
 // Indexer calculates the secondary index key for a given object
 type Indexer func(Object) ([]byte, error)
@@ -15,11 +41,11 @@ type Indexer func(Object) ([]byte, error)
 // MultiKeyIndexer calculates the secondary index keys for a given object
 type MultiKeyIndexer func(Object) ([][]byte, error)
 
-// Index represents a secondary index on some data.
+// compactIndex represents a secondary index on some data.
 // It is indexed by an arbitrary key returned by Indexer.
 // The value is one primary key (unique),
 // Or an array of primary keys (!unique).
-type Index struct {
+type compactIndex struct {
 	name   string
 	id     []byte
 	unique bool
@@ -27,27 +53,17 @@ type Index struct {
 	refKey func([]byte) []byte
 }
 
-var _ weave.QueryHandler = Index{}
-
-// NewIndex constructs an index with single key Indexer.
-// Indexer calculates the index for an object
-// unique enforces a unique constraint on the index
-// refKey calculates the absolute dbkey for a ref
-func NewIndex(name string, indexer Indexer, unique bool,
-	refKey func([]byte) []byte) Index {
-	return NewMultiKeyIndex(name, asMultiKeyIndexer(indexer), unique, refKey)
-}
+var _ weave.QueryHandler = compactIndex{}
 
 // NewMultiKeyIndex constructs an index with multi key indexer.
 // Indexer calculates the index for an object
 // unique enforces a unique constraint on the index
 // refKey calculates the absolute dbkey for a ref
-func NewMultiKeyIndex(name string, indexer MultiKeyIndexer, unique bool,
-	refKey func([]byte) []byte) Index {
+func NewMultiKeyIndex(name string, indexer MultiKeyIndexer, unique bool, refKey func([]byte) []byte) Index {
 	// TODO: index name must be [a-z_]
-	return Index{
+	return compactIndex{
 		name:   name,
-		id:     append(indPrefix, []byte(name+":")...),
+		id:     append([]byte(compactIdxPrefix), []byte(name+":")...),
 		index:  indexer,
 		unique: unique,
 		refKey: refKey,
@@ -67,10 +83,14 @@ func asMultiKeyIndexer(indexer Indexer) MultiKeyIndexer {
 	}
 }
 
-// IndexKey is the full key we store in the db, including prefix
+func (i compactIndex) Name() string {
+	return i.name
+}
+
+// indexKey is the full key we store in the db, including prefix
 // We copy into a new array rather than use append, as we don't
 // want consecutive calls to overwrite the same byte array.
-func (i Index) IndexKey(key []byte) []byte {
+func (i compactIndex) indexKey(key []byte) []byte {
 	l := len(i.id)
 	out := make([]byte, l+len(key))
 	copy(out, i.id)
@@ -88,7 +108,7 @@ func (i Index) IndexKey(key []byte) []byte {
 //
 // Otherwise, it will check indexer(prev) and indexer(save)
 // and make sure the key is now stored in the right location
-func (i Index) Update(db weave.KVStore, prev Object, save Object) error {
+func (i compactIndex) Update(db weave.KVStore, prev Object, save Object) error {
 	type s struct{ a, b bool }
 	sw := s{prev == nil, save == nil}
 	switch sw {
@@ -122,16 +142,16 @@ func (i Index) Update(db weave.KVStore, prev Object, save Object) error {
 	return errors.Wrap(errors.ErrHuman, "you have violated the rules of boolean logic")
 }
 
-// GetLike calculates the index for the given pattern, and
+// Like calculates the index for the given pattern, and
 // returns a list of all pk that match (may be nil when empty), or an error
-func (i Index) GetLike(db weave.ReadOnlyKVStore, pattern Object) ([][]byte, error) {
+func (i compactIndex) Like(db weave.ReadOnlyKVStore, pattern Object) ([][]byte, error) {
 	indexes, err := i.index(pattern)
 	if err != nil {
 		return nil, err
 	}
 	var r [][]byte
 	for _, index := range indexes {
-		pks, err := i.GetAt(db, index)
+		pks, err := consumeIteratorKeys(i.Keys(db, index))
 		if err != nil {
 			return nil, err
 		}
@@ -154,31 +174,80 @@ func deduplicate(s [][]byte) [][]byte {
 	return s
 }
 
-// GetAt returns a list of all pk at that index (may be empty), or an error
-func (i Index) GetAt(db weave.ReadOnlyKVStore, index []byte) ([][]byte, error) {
-	key := i.IndexKey(index)
+// Keys returns a list of all entity keys that were indexed under given value.
+func (i compactIndex) Keys(db weave.ReadOnlyKVStore, index []byte) weave.Iterator {
+	key := i.indexKey(index)
 	val, err := db.Get(key)
 	if err != nil {
-		return nil, err
+		return &failedIterator{err: err}
 	}
 	if val == nil {
-		return nil, nil
+		return &failedIterator{err: errors.ErrIteratorDone}
 	}
 	if i.unique {
-		return [][]byte{val}, nil
+		return &keysIterator{keys: [][]byte{val}}
 	}
-	var data = new(MultiRef)
-	err = data.Unmarshal(val)
-	if err != nil {
-		return nil, err
+
+	var data MultiRef
+	if err := data.Unmarshal(val); err != nil {
+		return &failedIterator{err: err}
 	}
-	return data.GetRefs(), nil
+	return &keysIterator{keys: data.GetRefs()}
 }
 
-// GetPrefix returns all references that have an index that
+type failedIterator struct {
+	err error
+}
+
+var _ weave.Iterator = (*failedIterator)(nil)
+
+func (it *failedIterator) Next() ([]byte, []byte, error) {
+	return nil, nil, it.err
+}
+
+func (failedIterator) Release() {}
+
+type keysIterator struct {
+	keys [][]byte
+}
+
+var _ weave.Iterator = (*keysIterator)(nil)
+
+func (it *keysIterator) Next() ([]byte, []byte, error) {
+	if len(it.keys) == 0 {
+		return nil, nil, errors.ErrIteratorDone
+	}
+	key := it.keys[0]
+	it.keys = it.keys[1:]
+	return key, nil, nil
+}
+
+func (keysIterator) Release() {}
+
+// consumeIteratorKeys returns a list of all keys that given iterator returns.
+// This function should be used only for iterators when the result size is
+// known to be small as all results are kept in memory.
+// This function releases the iterator.
+func consumeIteratorKeys(it weave.Iterator) ([][]byte, error) {
+	defer it.Release()
+
+	var keys [][]byte
+	for {
+		switch k, _, err := it.Next(); {
+		case err == nil:
+			keys = append(keys, k)
+		case errors.ErrIteratorDone.Is(err):
+			return keys, nil
+		default:
+			return keys, err
+		}
+	}
+}
+
+// getPrefix returns all references that have an index that
 // begins with a given prefix
-func (i Index) GetPrefix(db weave.ReadOnlyKVStore, prefix []byte) ([][]byte, error) {
-	dbPrefix := i.IndexKey(prefix)
+func (i compactIndex) getPrefix(db weave.ReadOnlyKVStore, prefix []byte) ([][]byte, error) {
+	dbPrefix := i.indexKey(prefix)
 	itr, err := db.Iterator(prefixRange(dbPrefix))
 	if err != nil {
 		return nil, err
@@ -207,18 +276,17 @@ func (i Index) GetPrefix(db weave.ReadOnlyKVStore, prefix []byte) ([][]byte, err
 }
 
 // Query handles queries from the QueryRouter
-func (i Index) Query(db weave.ReadOnlyKVStore, mod string,
-	data []byte) ([]weave.Model, error) {
+func (i compactIndex) Query(db weave.ReadOnlyKVStore, mod string, data []byte) ([]weave.Model, error) {
 
 	switch mod {
 	case weave.KeyQueryMod:
-		refs, err := i.GetAt(db, data)
+		refs, err := consumeIteratorKeys(i.Keys(db, data))
 		if err != nil {
 			return nil, err
 		}
 		return i.loadRefs(db, refs)
 	case weave.PrefixQueryMod:
-		refs, err := i.GetPrefix(db, data)
+		refs, err := i.getPrefix(db, data)
 		if err != nil {
 			return nil, err
 		}
@@ -228,7 +296,7 @@ func (i Index) Query(db weave.ReadOnlyKVStore, mod string,
 	}
 }
 
-func (i Index) loadRefs(db weave.ReadOnlyKVStore,
+func (i compactIndex) loadRefs(db weave.ReadOnlyKVStore,
 	refs [][]byte) ([]weave.Model, error) {
 
 	if len(refs) == 0 {
@@ -249,7 +317,7 @@ func (i Index) loadRefs(db weave.ReadOnlyKVStore,
 	return res, nil
 }
 
-func (i Index) move(db weave.KVStore, prev Object, save Object) error {
+func (i compactIndex) move(db weave.KVStore, prev Object, save Object) error {
 	// if the primary key is not equal, we have a problem
 	if !bytes.Equal(prev.Key(), save.Key()) {
 		return errors.Wrap(errors.ErrImmutable, "cannot modify the primary key of an object")
@@ -269,7 +337,7 @@ func (i Index) move(db weave.KVStore, prev Object, save Object) error {
 	// check unique constraints first
 	for _, newKey := range keysToAdd {
 		if i.unique {
-			k := i.IndexKey(newKey)
+			k := i.indexKey(newKey)
 			val, err := db.Get(k)
 			if err != nil {
 				return err
@@ -314,13 +382,13 @@ OUTER:
 	return r
 }
 
-func (i Index) remove(db weave.KVStore, index []byte, pk []byte) error {
+func (i compactIndex) remove(db weave.KVStore, index []byte, pk []byte) error {
 	// don't deal with empty keys
 	if len(index) == 0 {
 		return nil
 	}
 
-	key := i.IndexKey(index)
+	key := i.indexKey(index)
 	cur, err := db.Get(key)
 	if err != nil {
 		return err
@@ -359,13 +427,13 @@ func (i Index) remove(db weave.KVStore, index []byte, pk []byte) error {
 	return db.Set(key, save)
 }
 
-func (i Index) insert(db weave.KVStore, index []byte, pk []byte) error {
+func (i compactIndex) insert(db weave.KVStore, index []byte, pk []byte) error {
 	// don't deal with empty keys
 	if len(index) == 0 {
 		return nil
 	}
 
-	key := i.IndexKey(index)
+	key := i.indexKey(index)
 	cur, err := db.Get(key)
 	if err != nil {
 		return err
@@ -399,4 +467,180 @@ func (i Index) insert(db weave.KVStore, index []byte, pk []byte) error {
 	}
 
 	return db.Set(key, save)
+}
+
+const nativeIdxPrefix = "_x."
+
+func NewNativeIndex(name string, indexer MultiKeyIndexer) Index {
+	return &nativeIndex{
+		name:    name,
+		indexer: indexer,
+	}
+}
+
+type nativeIndex struct {
+	name    string
+	indexer MultiKeyIndexer
+}
+
+func (ix *nativeIndex) Name() string {
+	return ix.name
+}
+
+// Update updates the index. It should be called when any of the bucket
+// entities has changed in the store.
+//
+// prev == nil means insert
+// next == nil means delete
+// both == nil is error
+// if both != nil and prev.Key() != next.Key() this is an error
+func (ix *nativeIndex) Update(db weave.KVStore, prev Object, next Object) error {
+	if next == nil && prev == nil {
+		return errors.Wrap(errors.ErrInput, "update requires at least one non-nil object")
+	}
+	if next != nil && prev != nil {
+		if !bytes.Equal(next.Key(), prev.Key()) {
+			return errors.Wrap(errors.ErrState, "previous key is not the same as the new one")
+		}
+	}
+
+	// Delete.
+	if prev != nil {
+		values, err := ix.indexer(prev)
+		if err != nil {
+			return errors.Wrap(err, "indexer")
+		}
+		for _, v := range values {
+			idxKey, err := packNativeIdxKey([][]byte{[]byte(ix.name), v, prev.Key()})
+			if err != nil {
+				return errors.Wrap(err, "build index key")
+			}
+			if err := db.Delete(idxKey); err != nil {
+				return errors.Wrap(err, "db delete")
+			}
+		}
+	}
+
+	// Insert.
+	if next != nil {
+		values, err := ix.indexer(next)
+		if err != nil {
+			return errors.Wrap(err, "indexer")
+		}
+		for _, v := range values {
+			idxKey, err := packNativeIdxKey([][]byte{[]byte(ix.name), v, next.Key()})
+			if err != nil {
+				return errors.Wrap(err, "build index key")
+			}
+			if err := db.Set(idxKey, nil); err != nil {
+				return errors.Wrap(err, "db set")
+			}
+		}
+	}
+
+	return nil
+}
+
+func (ix *nativeIndex) Keys(db weave.ReadOnlyKVStore, value []byte) weave.Iterator {
+	lookupKey, err := packNativeIdxKey([][]byte{[]byte(ix.name), value})
+	if err != nil {
+		return &failedIterator{err: errors.Wrap(err, "build index key")}
+	}
+
+	// Index key are built is a specific way, that allow using the native
+	// database key iteration in order to find all indexed entries. Index
+	// key is in format:
+	//    <prefix>#<index name>#<value>#<entity id>
+	// where # is a serialization specific data, irrelevant for the
+	// algorithm.
+	// To iterate over all values matching given index, iterate over all
+	// keys between:
+	//    <prefix>#<index name>#<value> and <prefix>#<index name>#<value>{255}
+	//
+	// Parse matching keys and return the last part of it, being the
+	// indexed entity.
+	// Value 255 is reserved to make sure no indexed key is matching it
+	// (see packNativeIdxKey function).
+
+	start := lookupKey
+	end := make([]byte, len(lookupKey)+1)
+	copy(end, lookupKey)
+	// MaxUint8 is not used by serializer so we can use it as the maximum
+	// value guard.
+	end[len(end)-1] = math.MaxUint8
+
+	it, err := db.Iterator(start, end)
+	if err != nil {
+		return &failedIterator{err: err}
+	}
+
+	return &nativeIndexIterator{dbit: it}
+}
+
+// nativeIndexIterator wraps a database iterator and parse results to provide
+// indexed entities keys. It provides an interface that returns only the
+// relevant data, hiding from the user native index implementation details.
+type nativeIndexIterator struct {
+	dbit weave.Iterator
+}
+
+func (it *nativeIndexIterator) Release() {
+	it.dbit.Release()
+}
+
+func (it *nativeIndexIterator) Next() ([]byte, []byte, error) {
+	key, _, err := it.dbit.Next()
+	if err != nil {
+		return key, nil, err
+	}
+	chunks, err := unpackNativeIdxKey(key)
+	if err != nil {
+		return nil, nil, errors.Wrap(err, "unpack native index key")
+	}
+	return chunks[len(chunks)-1], nil, nil
+}
+
+func (ix *nativeIndex) Query(db weave.ReadOnlyKVStore, mod string, data []byte) ([]weave.Model, error) {
+	panic("not implemented")
+}
+
+func packNativeIdxKey(chunks [][]byte) ([]byte, error) {
+	var size int
+	for _, b := range chunks {
+		size += len(b) + 1
+	}
+	// First bytes are prefix information.
+	res := make([]byte, 0, size+len(nativeIdxPrefix))
+	res = append(res, nativeIdxPrefix...)
+
+	for _, b := range chunks {
+		// MaxUint8 is reserved for the search purpose. MaxUint8 - 1 is
+		// the greatest allowed length.
+		if len(b) > math.MaxUint8-1 {
+			return nil, errors.Wrapf(errors.ErrInput, "no chunk can be bigger than %d bytes", math.MaxUint8-1)
+		}
+		res = append(res, uint8(len(b)))
+		res = append(res, b...)
+	}
+	return res, nil
+}
+
+func unpackNativeIdxKey(b []byte) ([][]byte, error) {
+	if len(b) < len(nativeIdxPrefix) {
+		return nil, errors.Wrap(errors.ErrInput, "not a native index key")
+	}
+	if !bytes.Equal(b[:len(nativeIdxPrefix)], []byte(nativeIdxPrefix)) {
+		return nil, errors.Wrap(errors.ErrInput, "not a native index key")
+	}
+	b = b[len(nativeIdxPrefix):]
+	res := make([][]byte, 0, 6)
+	for len(b) > 0 {
+		size := uint8(b[0])
+		if len(b) < 1+int(size) {
+			return nil, errors.Wrap(errors.ErrInput, "malformed offset")
+		}
+		res = append(res, b[1:1+size])
+		b = b[1+size:]
+	}
+	return res, nil
 }