+++ orm/index_test.go
@@ -2,30 +2,41 @@ package orm
 
 import (
 	"bytes"
-	"errors"
+	stderrors "errors"
 	"fmt"
+	"reflect"
 	"testing"
 
+	"github.com/iov-one/weave"
+	"github.com/iov-one/weave/errors"
 	"github.com/iov-one/weave/store"
 	"github.com/iov-one/weave/weavetest/assert"
 )
 
+// newIndex constructs an index with single key Indexer.
+// Indexer calculates the index for an object
+// unique enforces a unique constraint on the index
+// refKey calculates the absolute dbkey for a ref
+func newIndex(name string, indexer Indexer, unique bool, refKey func([]byte) []byte) Index {
+	return NewMultiKeyIndex(name, asMultiKeyIndexer(indexer), unique, refKey)
+}
+
 // simple indexer for Counter
 func count(obj Object) ([]byte, error) {
 	if obj == nil {
-		return nil, errors.New("Cannot take index of nil")
+		return nil, stderrors.New("cannot take index of nil")
 	}
 	cntr, ok := obj.Value().(*Counter)
 	if !ok {
-		return nil, errors.New("Can only take index of Counter")
+		return nil, stderrors.New("can only take index of Counter")
 	}
 	// big-endian encoded int64
 	return encodeSequence(cntr.Count), nil
 }
 
 func TestCounterSingleKeyIndex(t *testing.T) {
-	multi := NewIndex("likes", count, false, nil)
-	uniq := NewIndex("magic", count, true, nil)
+	multi := newIndex("likes", count, false, nil).(compactIndex)
+	uniq := newIndex("magic", count, true, nil).(compactIndex)
 
 	// some keys to use
 	k1 := []byte("abc")
@@ -44,7 +55,7 @@ func TestCounterSingleKeyIndex(t *testing.T) {
 	e9 := encodeSequence(9)
 
 	cases := []struct {
-		idx        Index
+		idx        compactIndex
 		prev, next Object // for Update
 		isError    bool   // check Update result
 		// if there was no error, and these are non-nil, try
@@ -102,12 +113,12 @@ func TestCounterSingleKeyIndex(t *testing.T) {
 
 			assert.Nil(t, err)
 			if tc.getLike != nil {
-				res, err := idx.GetLike(db, tc.getLike)
+				res, err := idx.Like(db, tc.getLike)
 				assert.Nil(t, err)
 				assert.Equal(t, tc.likeRes, res)
 			}
 			if tc.getAt != nil {
-				res, err := idx.GetAt(db, tc.getAt)
+				res, err := consumeIteratorKeys(idx.Keys(db, tc.getAt))
 				assert.Nil(t, err)
 				assert.Equal(t, tc.atRes, res)
 			}
@@ -116,10 +127,10 @@ func TestCounterSingleKeyIndex(t *testing.T) {
 }
 
 func TestCounterMultiKeyIndex(t *testing.T) {
-	uniq := NewMultiKeyIndex("unique", evenOddIndexer, true, nil)
+	uniq := NewMultiKeyIndex("unique", evenOddIndexer, true, nil).(compactIndex)
 
 	specs := map[string]struct {
-		index               Index
+		index               compactIndex
 		store               Object
 		prev, next          Object
 		expError            bool
@@ -157,7 +168,7 @@ func TestCounterMultiKeyIndex(t *testing.T) {
 			expError: true,
 		},
 		"update without unique constraint": {
-			index:    NewMultiKeyIndex("multi", evenOddIndexer, false, nil),
+			index:    NewMultiKeyIndex("multi", evenOddIndexer, false, nil).(compactIndex),
 			store:    NewSimpleObj([]byte("even"), NewCounter(8)),
 			prev:     NewSimpleObj([]byte("my"), NewCounter(5)),
 			next:     NewSimpleObj([]byte("my"), NewCounter(6)),
@@ -202,7 +213,7 @@ func TestCounterMultiKeyIndex(t *testing.T) {
 			}
 			for _, k := range spec.expKeys {
 				// and index keys exists
-				pks, err := idx.GetAt(db, k)
+				pks, err := consumeIteratorKeys(idx.Keys(db, k))
 				assert.Nil(t, err)
 				// with proper pk
 				if idx.unique {
@@ -220,7 +231,7 @@ func TestCounterMultiKeyIndex(t *testing.T) {
 			}
 			// and previous index keys don't exist anymore
 			for _, k := range spec.expNotKeys {
-				pks, err := idx.GetAt(db, k)
+				pks, err := consumeIteratorKeys(idx.Keys(db, k))
 				assert.Nil(t, err)
 				assert.Nil(t, pks)
 			}
@@ -230,7 +241,7 @@ func TestCounterMultiKeyIndex(t *testing.T) {
 
 func TestGetLikeWithMultiKeyIndex(t *testing.T) {
 	db := store.MemStore()
-	idx := NewMultiKeyIndex("multi", evenOddIndexer, false, nil)
+	idx := NewMultiKeyIndex("multi", evenOddIndexer, false, nil).(compactIndex)
 
 	persistentObjects := []Object{
 		NewSimpleObj([]byte("firstOdd"), NewCounter(5)),
@@ -267,7 +278,7 @@ func TestGetLikeWithMultiKeyIndex(t *testing.T) {
 	}
 	for testName, spec := range specs {
 		t.Run(testName, func(t *testing.T) {
-			pks, err := idx.GetLike(db, spec.source)
+			pks, err := idx.Like(db, spec.source)
 			// then
 			assert.Nil(t, err)
 			assert.Equal(t, spec.expPKs, pks)
@@ -292,11 +303,11 @@ func evenOddIndexer(obj Object) ([][]byte, error) {
 // return first value (if any), or nil
 func first(obj Object) ([]byte, error) {
 	if obj == nil {
-		return nil, errors.New("Cannot take index of nil")
+		return nil, stderrors.New("Cannot take index of nil")
 	}
 	multi, ok := obj.Value().(*MultiRef)
 	if !ok {
-		return nil, errors.New("Can only take index of MultiRef")
+		return nil, stderrors.New("Can only take index of MultiRef")
 	}
 	if len(multi.Refs) == 0 {
 		return nil, nil
@@ -366,7 +377,7 @@ func TestNullableIndex(t *testing.T) {
 
 	for testName, tc := range cases {
 		t.Run(testName, func(t *testing.T) {
-			uniq := NewIndex("no-null", first, true, nil)
+			uniq := newIndex("no-null", first, true, nil)
 			db := store.MemStore()
 			for _, init := range tc.setup {
 				err := uniq.Update(db, nil, init)
@@ -430,4 +441,217 @@ func toBytes(s []string) [][]byte {
 		source[i] = []byte(v)
 	}
 	return source
+}
+
+func TestNativeIndexPacking(t *testing.T) {
+	cases := map[string][][]byte{
+		"empty":                [][]byte{},
+		"one empty element":    [][]byte{[]byte{}},
+		"three empty elements": [][]byte{[]byte{}, []byte{}, []byte{}},
+		"one non empty element": [][]byte{
+			[]byte("foo"),
+		},
+		"two non empty elements": [][]byte{
+			[]byte("a"),
+			[]byte("a very long value that is below 255 characters"),
+		},
+		"three non empty elements": [][]byte{
+			[]byte("foo"),
+			[]byte("bar"),
+			[]byte("baz"),
+		},
+		"mixture of empty and non empty": [][]byte{
+			[]byte("non empty value"),
+			[]byte{},
+			[]byte("another non empty value"),
+			[]byte{},
+			[]byte{},
+			[]byte{},
+			[]byte("not empty"),
+			[]byte{},
+		},
+	}
+
+	for testName, chunks := range cases {
+		t.Run(testName, func(t *testing.T) {
+			packed, err := packNativeIdxKey(chunks)
+			if err != nil {
+				t.Fatalf("cannot pack: %s", err)
+			}
+			unpacked, err := unpackNativeIdxKey(packed)
+			if err != nil {
+				t.Fatalf("cannot unpack: %s", err)
+			}
+			if !reflect.DeepEqual(unpacked, chunks) {
+				t.Logf("packed: %x %q", packed, packed)
+				t.Fatalf("data malformed during serialization: %q", unpacked)
+			}
+		})
+	}
+}
+
+func TestCompactIndexImplementation(t *testing.T) {
+	testIndexImplementation(t, func(fn MultiKeyIndexer) Index {
+		return NewMultiKeyIndex("myindex", fn, false, func(b []byte) []byte { return b })
+	})
+}
+
+func TestNativeIndexImplementation(t *testing.T) {
+	testIndexImplementation(t, func(fn MultiKeyIndexer) Index {
+		return NewNativeIndex("myindex", fn)
+	})
+}
+
+func testIndexImplementation(t *testing.T, newIdx func(MultiKeyIndexer) Index) {
+	valueIdx := func(o Object) ([][]byte, error) {
+		c := o.Value().(*Counter).Count
+		return [][]byte{[]byte(fmt.Sprint(c))}, nil
+	}
+	idx := newIdx(valueIdx)
+
+	// Definition of a single Update method call and expected result.
+	type updateCall struct {
+		prev    Object
+		next    Object
+		wantErr *errors.Error
+	}
+
+	// Definition of a single Keys method call and expected results.
+	type keysCall struct {
+		value    []byte
+		wantKeys []string // []string and not [][]byte for nicer UI
+	}
+
+	cases := map[string]struct {
+		idx     Index
+		updates []updateCall
+		keys    []keysCall
+	}{
+		"no results found": {
+			idx:     idx,
+			updates: []updateCall{},
+			keys: []keysCall{
+				{value: []byte("random-value"), wantKeys: nil},
+			},
+		},
+		"a single item": {
+			idx: idx,
+			updates: []updateCall{
+				{prev: nil, next: NewSimpleObj([]byte("one"), &Counter{Count: 1})},
+			},
+			keys: []keysCall{
+				{value: []byte("1"), wantKeys: []string{"one"}},
+				{value: []byte("unindexed-value"), wantKeys: nil},
+			},
+		},
+		"two items, both with the same index value": {
+			idx: idx,
+			updates: []updateCall{
+				{prev: nil, next: NewSimpleObj([]byte("first"), &Counter{Count: 1})},
+				{prev: nil, next: NewSimpleObj([]byte("second"), &Counter{Count: 1})},
+			},
+			keys: []keysCall{
+				{value: []byte("1"), wantKeys: []string{"first", "second"}},
+				{value: []byte("unindexed-value"), wantKeys: nil},
+			},
+		},
+		"two items, each with a different index value": {
+			idx: idx,
+			updates: []updateCall{
+				{prev: nil, next: NewSimpleObj([]byte("one"), &Counter{Count: 1})},
+				{prev: nil, next: NewSimpleObj([]byte("two"), &Counter{Count: 2})},
+			},
+			keys: []keysCall{
+				{value: []byte("1"), wantKeys: []string{"one"}},
+				{value: []byte("2"), wantKeys: []string{"two"}},
+				{value: []byte("unindexed-value"), wantKeys: nil},
+			},
+		},
+		"many items, some with similar index value": {
+			idx: idx,
+			updates: []updateCall{
+				{prev: nil, next: NewSimpleObj([]byte("a"), &Counter{Count: 1})},
+				{prev: nil, next: NewSimpleObj([]byte("b"), &Counter{Count: 2})},
+				{prev: nil, next: NewSimpleObj([]byte("c"), &Counter{Count: 2})},
+				{prev: nil, next: NewSimpleObj([]byte("d"), &Counter{Count: 2})},
+				{prev: nil, next: NewSimpleObj([]byte("e"), &Counter{Count: 3})},
+			},
+			keys: []keysCall{
+				{value: []byte("1"), wantKeys: []string{"a"}},
+				{value: []byte("2"), wantKeys: []string{"b", "c", "d"}},
+				{value: []byte("3"), wantKeys: []string{"e"}},
+				{value: []byte("unindexed-value"), wantKeys: nil},
+			},
+		},
+		"deleting an item from an index": {
+			idx: idx,
+			updates: []updateCall{
+				{prev: nil, next: NewSimpleObj([]byte("one"), &Counter{Count: 1})},
+				{prev: nil, next: NewSimpleObj([]byte("two"), &Counter{Count: 2})},
+				{prev: NewSimpleObj([]byte("two"), &Counter{Count: 2}), next: nil},
+			},
+			keys: []keysCall{
+				{value: []byte("1"), wantKeys: []string{"one"}},
+				{value: []byte("2"), wantKeys: nil},
+			},
+		},
+		"reindexing an item with a different value": {
+			idx: idx,
+			updates: []updateCall{
+				{prev: nil, next: NewSimpleObj([]byte("two"), &Counter{Count: 1})},
+				{prev: nil, next: NewSimpleObj([]byte("one"), &Counter{Count: 1})},
+				{prev: NewSimpleObj([]byte("two"), &Counter{Count: 1}), next: NewSimpleObj([]byte("two"), &Counter{Count: 2})},
+			},
+			keys: []keysCall{
+				{value: []byte("1"), wantKeys: []string{"one"}},
+				{value: []byte("2"), wantKeys: []string{"two"}},
+				{value: []byte("unindexed-value"), wantKeys: nil},
+			},
+		},
+	}
+
+	for testName, tc := range cases {
+		t.Run(testName, func(t *testing.T) {
+			db := store.MemStore()
+
+			for i, u := range tc.updates {
+				if err := tc.idx.Update(db, u.prev, u.next); !u.wantErr.Is(err) {
+					t.Fatalf("%d update: want %q error, got %q", i, u.wantErr, err)
+				}
+			}
+
+			for i, k := range tc.keys {
+				keys, err := iteratorKeys(tc.idx.Keys(db, k.value))
+				if err != nil {
+					t.Fatalf("%d iterator keys failed: %s", i, err)
+				}
+
+				var want [][]byte
+				for _, k := range k.wantKeys {
+					want = append(want, []byte(k))
+				}
+				if !reflect.DeepEqual(keys, want) {
+					t.Logf("want keys: %q", want)
+					t.Logf(" got keys: %q", keys)
+					t.Errorf("%d keys call returned unexpected keys for value %q", i, k.value)
+				}
+			}
+		})
+	}
+}
+
+func iteratorKeys(it weave.Iterator) ([][]byte, error) {
+	defer it.Release()
+
+	var res [][]byte
+	for {
+		switch key, _, err := it.Next(); {
+		case err == nil:
+			res = append(res, key)
+		case errors.ErrIteratorDone.Is(err):
+			return res, nil
+		default:
+			return res, err
+		}
+	}
 }